+++ src/Exception/CacheMissingConfigurationException.php
@@ -8,12 +8,8 @@ use Throwable;
 /**
  *
  * Exception thrown when some mandatory configurations are missing during cache creation
- * @author Damiano Improta <code@damianoimprota.dev>
+ * @author Damiano Improta <code@damianoimprota.it>
  */
 class CacheMissingConfigurationException extends Exception
 {
-    public function __construct(string $message = "", int $code = 0, ?Throwable $previous = null)
-    {
-        parent::__construct($message, $code, $previous);
-    }
 }