+++ tests/ApcuCacheTest.php
@@ -11,7 +11,7 @@ use Override;
 /**
  * APCU unit test class implementation
  *
- * @author Damiano Improta <code@damianoimprota.dev> 
+ * @author Damiano Improta <code@damianoimprota.it> 
  */
 class ApcuCacheTest extends AbstractCache
 {
@@ -143,6 +143,12 @@ class ApcuCacheTest extends AbstractCache
     {
         parent::tearDownAfterClass();
         Cache::factory(CacheEnum::APCU, 60)->clearAllCache();
+    }
+    
+    #[\Override]
+    public function testArrayDepth(): void
+    {
+        parent::testArrayDepth();
     }
 
     /**