+++ tests/Entity/Foo.php
@@ -9,7 +9,7 @@ use Override;
  * 
  * test class, used only for the phpunit test suite
  * 
- * @author Damiano Improta <code@damianoimprota.dev> 
+ * @author Damiano Improta <code@damianoimprota.it> 
  */
 class Foo implements Cacheable {