+++ tests/MemcacheCacheTest.php
@@ -6,12 +6,13 @@ use CacheMultiLayer\Enum\CacheEnum;
 use CacheMultiLayer\Exception\CacheMissingConfigurationException;
 use CacheMultiLayer\Service\Cache;
 use Exception;
+use Memcache;
 use Override;
 
 /**
  * MEMCACHE unit test class implementation
  *
- * @author Damiano Improta <code@damianoimprota.dev> 
+ * @author Damiano Improta <code@damianoimprota.it> 
  */
 class MemcacheCacheTest extends AbstractCache
 {
@@ -26,7 +27,7 @@ class MemcacheCacheTest extends AbstractCache
     #[Override]
     public static function setUpBeforeClass(): void
     {
-       parent::setUpBeforeClass();
+        parent::setUpBeforeClass();
         set_error_handler(function (int $errno, string $errstr, string $errfile, int $errline): bool {
             // error was suppressed with the @-operator
             if (0 === error_reporting()) {
@@ -101,7 +102,7 @@ class MemcacheCacheTest extends AbstractCache
     {
         parent::testIsConnected();
     }
-    
+
     #[\Override]
     public function testEmptyDecrement(): void
     {
@@ -132,15 +133,46 @@ class MemcacheCacheTest extends AbstractCache
         Cache::factory(CacheEnum::MEMCACHE, 60, ['server_address' => 'ip-no-memcache'])->isConnected();
     }
 
+    public function testInstance(): void
+    {
+        $memcache = new Memcache();
+        $memcache->connect('memcache-server', 11211);
+        Cache::factory(CacheEnum::MEMCACHE, 60, ['instance' => $memcache]);
+        $this->assertTrue(true); //no exception throwns
+    }
+
+    public function testMissingInstance(): void
+    {
+        $this->expectException(CacheMissingConfigurationException::class);
+        Cache::factory(CacheEnum::MEMCACHE, 60, ['instance' => 5]);
+    }
+
     public function testEnum(): void
     {
         $this->doTestRealEnum(CacheEnum::MEMCACHE);
     }
 
+    public function testPrefix(): void
+    {
+        $val = 10; //maradona
+        $key = "test_prefix";
+        $cacheSamePrefix = Cache::factory(CacheEnum::MEMCACHE, 60, ['key_prefix' => '', 'server_address' => 'memcache-server']);
+        $cacheOtherPrefix = Cache::factory(CacheEnum::MEMCACHE, 10, ['key_prefix' => 'other_', 'server_address' => 'memcache-server']);
+        $this->getCache()->set($key, $val);
+        $this->assertEquals($cacheSamePrefix->get($key), $val);
+        $this->assertNull($cacheOtherPrefix->get($key));
+    }
+
+    #[\Override]
+    public function testArrayDepth(): void
+    {
+        parent::testArrayDepth();
+    }
+
     #[\Override]
     public static function tearDownAfterClass(): void
     {
-         restore_error_handler();
+        restore_error_handler();
         Cache::factory(CacheEnum::MEMCACHE, 60, ['server_address' => 'memcache-server'])->clearAllCache();
     }
 }